# ERP/docs/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: ERP Main APIs for Apifox
  version: "1.0.0"
  description: |
    Main APIs extracted from current backend controllers.
    Authentication: all endpoints require Bearer token except /user/login.
servers:
  - url: http://localhost:8090
...
paths:
...
  /purchase/listPageC1:
    post:
      tags: [Purchase]
      summary: Purchase orders page query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryPageParamPurchase"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /purchase/detail:
    get:
      tags: [Purchase]
      summary: Purchase order detail
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /purchase/save:
    post:
      tags: [Purchase]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderDto"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /purchase/update:
    post:
      tags: [Purchase]
      summary: Update purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderDto"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /purchase/approve:
    post:
      tags: [Purchase]
      summary: Approve purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequest"
            example:
              id: 101
              operatorId: 1
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /purchase/del:
    delete:
      tags: [Purchase]
      summary: Delete purchase order
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /sales/listPageC1:
    post:
      tags: [Sales]
      summary: Sales orders page query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryPageParamSales"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /sales/detail:
    get:
      tags: [Sales]
      summary: Sales order detail
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /sales/save:
    post:
      tags: [Sales]
      summary: Create sales order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesOrderDto"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /sales/update:
    post:
      tags: [Sales]
      summary: Update sales order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesOrderDto"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /sales/approve:
    post:
      tags: [Sales]
      summary: Approve sales order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequest"
            example:
              id: 201
              operatorId: 1
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /sales/del:
    delete:
      tags: [Sales]
      summary: Delete sales order
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Result:
      type: object
      description: Loose generic response wrapper for compatibility in Apifox validation.
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: success
        total:
          type: integer
          format: int64
          example: 0
        data:
          description: Generic business payload, shape varies by endpoint and may be null.
          example: null
...
    QueryPageParamPurchase:
      type: object
      properties:
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        param:
          type: object
          properties:
            keyword:
              type: string
            status:
              type: string
              enum: [draft, pending, completed, cancelled]
    QueryPageParamSales:
      type: object
      properties:
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        param:
          type: object
          properties:
            keyword:
              type: string
            status:
              type: string
              enum: [draft, pending, completed, cancelled]
...
    ApproveRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        operatorId:
          type: integer
      required: [id, operatorId]
...
    PurchaseOrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        lineNo:
          type: integer
        productCode:
          type: string
        productName:
          type: string
        specification:
          type: string
        unit:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        amount:
          type: number
    PurchaseOrderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNo:
          type: string
        orderDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, pending, completed, cancelled]
        supplierName:
          type: string
        supplierContact:
          type: string
        supplierPhone:
          type: string
        supplierAddress:
          type: string
        totalQuantity:
          type: number
        totalAmount:
          type: number
        remark:
          type: string
        operatorId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
      required: [orderNo, orderDate, status, supplierName, supplierAddress, items]
    SalesOrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        lineNo:
          type: integer
        productCode:
          type: string
        productName:
          type: string
        specification:
          type: string
        unit:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        amount:
          type: number
    SalesOrderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNo:
          type: string
        orderDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, pending, completed, cancelled]
        customerName:
          type: string
        customerContact:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        itemCount:
          type: integer
        totalQuantity:
          type: number
        totalAmount:
          type: number
        remark:
          type: string
        operatorId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/SalesOrderItem"
      required:
        [orderNo, orderDate, status, customerName, customerContact, customerPhone, customerAddress, items]
...
security:
  - bearerAuth: []